requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
python-dotenv>=1.0.0
sqlalchemy>=2.0.0
pillow>=10.0.0
pillow-heif>=0.13.0
numpy>=1.24.0
google-generativeai>=0.8.0
st-supabase-connection
supabase>=2.0.0
bcrypt>=4.0.0
streamlit-option-menu
extra-streamlit-components
fpdf2>=2.7.0
